スペースを入れるoverrideを追加してスタンプとチャンネルの補完時にスペースがつくように
