+++ src/components/Main/MainView/MessageInput/composables/suggestion/overrides/stampSuggestion.ts
@@ -1,11 +1,6 @@
-import { type MaybeRefOrGetter, type Ref, computed, toValue } from 'vue'
+import { type MaybeRefOrGetter, computed, toValue } from 'vue'
 
-import type {
-  Candidate,
-  Target,
-  Word,
-  WordWithId
-} from '/@/lib/suggestion/basic'
+import type { Candidate, Word, WordWithId } from '/@/lib/suggestion/basic'
 import { useStampHistory } from '/@/store/domain/stampHistory'
 import { useStampRecommendations } from '/@/store/domain/stampRecommendations'
 
@@ -14,8 +9,6 @@ const stampSuggestionOverride = <
     onSelect: (word: Word) => void
     confirmedPart: MaybeRefOrGetter<string>
     suggestedCandidates: MaybeRefOrGetter<Candidate[]>
-    textareaRef: MaybeRefOrGetter<HTMLTextAreaElement | undefined>
-    target: Ref<Target>
   }
 >(
   input: Params
@@ -37,25 +30,10 @@ const stampSuggestionOverride = <
   })
 
   const onSelect = (word: WordWithId) => {
-    const textarea = toValue(input.textareaRef)
-    const target = input.target
-    if (!textarea) return
-    let insertingText: string
-    if (!textarea.value || target.value.begin === 0) {
-      insertingText = `${word.text}: `
-    } else {
-      const prevChar = textarea.value[target.value.begin - 1]
-      if (prevChar === ' ' || prevChar === '\n') {
-        insertingText = `${word.text}: `
-      } else {
-        insertingText = ` ${word.text}: `
-      }
-    }
-    
     upsertLocalStampHistory(word.id, new Date())
     recordStampUsage(word.id)
 
-    input.onSelect({ ...word, text: insertingText })
+    input.onSelect({ ...word, text: `${word.text}:` })
   }
 
   return {
